// src/generated/model/formueFrontend.ts
/**
 * Generated by orval v6.11.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface FormueFrontend {
    brukskonto: boolean;
    sparekonto: boolean;
    bsu: boolean;
    livsforsikring: boolean;
    verdipapirer: boolean;
    annet: boolean;
    beskrivelseAvAnnet?: string;
}